+++ src/requirements.txt
@@ -6,3 +6,4 @@ pydantic==2.7.0
 pymongo==4.6.2
 pytest==8.3.5
 python-dotenv==1.0.1
+fastapi_websocket_rpc==0.1.29
