+++ src/services/order_service.py
@@ -1,14 +1,11 @@
-import uuid
 from domain.order import Order
 from domain.order_repository_interface import OrderRepository
 from domain.product import Product
 from domain.product_repository_interface import ProductRepository
-from domain.user_repository_interface import UserRepository
 
 
 class OrderService:
-    def __init__(self, user_repo: UserRepository, product_repo: ProductRepository, order_repo: OrderRepository):
-        self.user_repo = user_repo
+    def __init__(self, product_repo: ProductRepository, order_repo: OrderRepository):
         self.product_repo = product_repo
         self.order_repo = order_repo
     def get_order(self, id: str) -> Order:
@@ -16,8 +13,7 @@ class OrderService:
             return self.order_repo.get(id)
         except:
             raise ValueError('Order not found')
-    def process_order(self, order: Order) -> Order:
-        user = self.user_repo.get(order.buyer_id)
+    def process_order(self,user, order: Order) -> Order:
         if user is None: raise ValueError("User not found")
         product = Product(**self.product_repo.get(order.product_id))
         if product is None: raise ValueError("Product not found")
